Rename useFirebaseLogin hook
